LayoutTests/media/media-play-promise-reject-error-notsupported-expected.txt


RUN(mediaElement.src = findMediaFile("video", "content/invalid"))
RUN(mediaElement.play().then(failTest).catch(promiseRejected))
Promise rejected. OK
EXPECTED (error.name == 'NotSupportedError') OK
END OF TEST
